+++ yandex/provider.go
@@ -250,6 +250,7 @@ func sdkProvider(emptyFolder bool) *schema.Provider {
 			"yandex_dataproc_cluster":                                 resourceYandexDataprocCluster(),
 			"yandex_datatransfer_endpoint":                            resourceYandexDatatransferEndpoint(),
 			"yandex_datatransfer_transfer":                            resourceYandexDatatransferTransfer(),
+			"yandex_dns_zone_iam_binding":                             resourceYandexDnsZoneIAMBinding(),
 			"yandex_dns_recordset":                                    resourceYandexDnsRecordSet(),
 			"yandex_dns_zone":                                         resourceYandexDnsZone(),
 			"yandex_function":                                         resourceYandexFunction(),