+++ yandex/resource_yandex_dns_zone_iam_binding.go
@@ -1,0 +1,18 @@
+package yandex
+
+import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+func resourceYandexDnsZoneIAMBinding() *schema.Resource {
+	return resourceIamBinding(
+		IamDnsZoneSchema,
+		newDnsZoneIamUpdater,
+		WithTimeout(
+			&schema.ResourceTimeout{
+				Default: schema.DefaultTimeout(yandexIAMDnsZoneDefaultTimeout),
+			}),
+		WithImporter(
+			&schema.ResourceImporter{
+				StateContext: iamBindingImport(dnsZoneIDParseFunc),
+			}),
+	)
+}